[INTERNAL] Move Reader and Transformer to 'readers' subdir
