+++ lib/AbstractReader.js
@@ -71,16 +71,16 @@ class AbstractReader {
 	}
 
 	filter(callback) {
-		const ReaderFilter = require("./ReaderFilter");
-		return new ReaderFilter({
+		const Filter = require("./readers/Filter");
+		return new Filter({
 			reader: this,
 			callback
 		});
 	}
 
 	transformer(callback) {
-		const ReaderTransformer = require("./ReaderTransformer");
-		return new ReaderTransformer({
+		const Transformer = require("./readers/Transformer");
+		return new Transformer({
 			reader: this,
 			callback
 		});
