+++ lib/readers/Filter.js
@@ -1,4 +1,4 @@
-const AbstractReader = require("./AbstractReader");
+const AbstractReader = require("../AbstractReader");
 
 /**
  * Transparently apply filters on resource readers by wrapping them.
@@ -7,12 +7,12 @@ const AbstractReader = require("./AbstractReader");
  * @memberof module:@ui5/fs
  * @augments module:@ui5/fs.AbstractReader
  */
-class ReaderFilter extends AbstractReader {
+class Filter extends AbstractReader {
 	/**
 	* Filter callback
 	*
 	* @public
-	* @callback module:@ui5/fs.ReaderFilter~callback
+	* @callback module:@ui5/fs.readers.Filter~callback
 	* @param {module:@ui5/fs.Resource} resource Resource to test
 	* @returns {boolean} Return <code>true</code> to keep the resource, or <code>false</code> to disregard it
 	*/
@@ -22,7 +22,7 @@ class ReaderFilter extends AbstractReader {
 	 *
 	 * @param {object} parameters Parameters
 	 * @param {module:@ui5/fs.AbstractReader} parameters.reader The resource reader to wrap
-	 * @param {module:@ui5/fs.ReaderFilter~callback} parameters.callback
+	 * @param {module:@ui5/fs.readers.Filter~callback} parameters.callback
 	 * 				Filter function. Will be called for every resource read through this reader.
 	 */
 	constructor({reader, callback}) {
@@ -72,4 +72,4 @@ class ReaderFilter extends AbstractReader {
 	}
 }
 
-module.exports = ReaderFilter;
+module.exports = Filter;
