+++ lib/readers/Transformer.js
@@ -1,18 +1,18 @@
-const AbstractReader = require("./AbstractReader");
+const AbstractReader = require("../AbstractReader");
 
 /**
  * Transparently apply filters on resource readers by wrapping them.
  *
  * @public
- * @memberof module:@ui5/fs
+ * @memberof module:@ui5/fs.readers
  * @augments module:@ui5/fs.AbstractReader
  */
-class ReaderTransformer extends AbstractReader {
+class Transformer extends AbstractReader {
 	/**
 	* Filter callback
 	*
 	* @public
-	* @callback module:@ui5/fs.ReaderTransformer~callback
+	* @callback module:@ui5/fs.readers.Transformer~callback
 	* @param {module:@ui5/fs.Resource} resource Resource to transform
 	* @returns {Promise|undefined} Optional promise resolving once the resource has been transformed
 	*/
@@ -22,7 +22,7 @@ class ReaderTransformer extends AbstractReader {
 	 *
 	 * @param {object} parameters Parameters
 	 * @param {module:@ui5/fs.AbstractReader} parameters.reader The resource reader to wrap
-	 * @param {module:@ui5/fs.ReaderTransformer~callback} parameters.callback
+	 * @param {module:@ui5/fs.readers.Transformer~callback} parameters.callback
 	 * 				Filter function. Will be called for every resource read through this reader.
 	 */
 	constructor({reader, callback}) {
@@ -81,4 +81,4 @@ class ReaderTransformer extends AbstractReader {
 	}
 }
 
-module.exports = ReaderTransformer;
+module.exports = Transformer;
